Integrate link in downloads page
