+++ apps/web-giddh/src/app/models/api-models/downloads.ts
@@ -14,3 +14,15 @@ export class DownloadsResponse {
     public items?: [];
 }
 
+export interface DownloadData {
+    date?: any;
+    requestedDate?: any;
+    user?: any;
+    type?: any;
+    filters?: any;
+    download?: any;
+    expireAt?: any;
+    requestId?: any;
+    status?: any
+}
+
