+++ apps/web-giddh/src/app/shared/header/components/account-operations/account-operations.component.ts
@@ -679,6 +679,7 @@ export class AccountOperationsComponent implements OnInit, AfterViewInit, OnDest
                 this.groupExportLedgerBodyRequest.sort = response.body.sort ? 'ASC' : 'DESC';
                 this.ledgerService.groupLedgerExport(this.groupExportLedgerBodyRequest).pipe(takeUntil(this.destroyed$)).subscribe(response => {
                     if (response.status === 'success') {
+                        this.toaster.showSnackBar("success", response.message);
                         this.router.navigate(["/pages/downloads"]);
                     } else {
                         this.toaster.showSnackBar("error", response.message);
